Add WrapFilter and implement AuthFilter with it
This commit introduces the notion of a "WrapFilter" which wraps a
(T, Req) => Resp service with a Req => Resp service using a lookup
function.

The most obvious use case for this is the AuthFilter. The lookup filter
looks at some cookie id, and returns the associated user, passing it to
the internal service. If there is no user, it returns Future.never (a
value that can be configured)